🚚 moved entities to features/person
