+++ libs/api/database/entities/src/lib/entities/features/index.ts
@@ -1,1 +1,2 @@
 // export * from './trip';
+export * from './person';
