+++ libs/api/database/entities/src/lib/entities/index.ts
@@ -1,3 +1,2 @@
 export * from './common';
-export * from './user';
-// export * from './features';
+export * from './features';
